# docs/swagger.yaml
basePath: /api
definitions:
...
  dto.AddEvent:
    properties:
      apps_downloaded:
        type: integer
      apps_name:
        type: string
      attendees_count:
        type: integer
      city_id:
        type: string
      description:
        type: string
      district_id:
        type: string
      end_time:
        type: string
      event_date:
        type: string
      event_type:
        type: string
      instructor_name:
        type: string
      instructor_phone:
        type: string
      location:
        type: string
      notes:
        type: string
      on_the_spot_sales:
        items:
          $ref: '#/definitions/dto.OnTheSpotSaleItem'
        type: array
      photos:
        items:
          $ref: '#/definitions/dto.AddEventPhoto'
        type: array
      province_id:
        type: string
      public_id:
        type: string
      school_id:
        type: string
      start_time:
        type: string
      status:
        type: string
      target_attendees:
        type: integer
      target_audience:
        type: string
      title:
        maxLength: 200
        minLength: 3
        type: string
      visiting_service_profit:
        type: number
      visiting_service_unit_entry:
        type: integer
    required:
    - city_id
    - description
    - district_id
    - end_time
    - event_date
    - event_type
    - location
    - province_id
    - start_time
    - title
    type: object
...
  dto.AddEventPhoto:
    properties:
      caption:
        type: string
      photo_order:
        type: integer
      photo_url:
        type: string
    required:
    - photo_url
    type: object
...
  dto.OnTheSpotSaleItem:
    properties:
      payment_method:
        type: string
      quantity:
        type: integer
      vehicle_type:
        type: string
    type: object
...
  dto.UpdateEvent:
    properties:
      apps_downloaded:
        type: integer
      apps_name:
        type: string
      attendees_count:
        type: integer
      city_id:
        type: string
      description:
        type: string
      district_id:
        type: string
      end_time:
        type: string
      event_date:
        type: string
      event_type:
        type: string
      instructor_name:
        type: string
      instructor_phone:
        type: string
      location:
        type: string
      notes:
        type: string
      on_the_spot_sales:
        items:
          $ref: '#/definitions/dto.OnTheSpotSaleItem'
        type: array
      province_id:
        type: string
      public_id:
        type: string
      school_id:
        type: string
      start_time:
        type: string
      status:
        type: string
      target_attendees:
        type: integer
      target_audience:
        type: string
      title:
        type: string
      visiting_service_profit:
        type: number
      visiting_service_unit_entry:
        type: integer
    type: object
...
  response.Error:
    properties:
      code:
        type: integer
      data: {}
      error: {}
      log_id:
        type: string
      message:
        type: string
      status:
        type: boolean
    type: object
  response.Success:
    properties:
      code:
        type: integer
      data: {}
      error: {}
      log_id:
        type: string
      message:
        type: string
      status:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: zaiduszhuhur@gmail.com
    name: API Support
    url: https://www.linkedin.com/in/zaidus-zhuhur/
  description: This is a sample server for a safety riding service.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Safety Riding API
  version: "1.0"
paths:
...
  /event:
    post:
      consumes:
      - application/json
      description: Create a new safety riding event
      parameters:
      - description: Event payload
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/dto.AddEvent'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - ApiKeyAuth: []
      summary: Create a new event
      tags:
      - Events
  /event/{id}:
    delete:
      consumes:
      - application/json
      description: Delete event by ID
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - ApiKeyAuth: []
      summary: Delete an event
      tags:
      - Events
    get:
      consumes:
      - application/json
      description: Retrieve event information by ID
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Success'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - ApiKeyAuth: []
      summary: Get event detail
      tags:
      - Events
    put:
      consumes:
      - application/json
      description: Update event information by ID
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      - description: Event payload
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateEvent'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - ApiKeyAuth: []
      summary: Update an event
      tags:
      - Events
  /event/{id}/photos:
    post:
      consumes:
      - multipart/form-data
      description: Upload one or more event photos and metadata
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      - description: Event photos
        in: formData
        name: photos
        required: true
        type: file
      - description: Photo captions (repeat per photo)
        in: formData
        name: captions
        type: string
      - description: Photo order (repeat per photo)
        in: formData
        name: photo_orders
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - ApiKeyAuth: []
      summary: Upload event photos
      tags:
      - Events
  /event/photo/{photoId}:
    delete:
      consumes:
      - application/json
      description: Delete a single event photo by ID
      parameters:
      - description: Photo ID
        in: path
        name: photoId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - ApiKeyAuth: []
      summary: Delete event photo
      tags:
      - Events
  /events:
    get:
      consumes:
      - application/json
      description: Retrieve paginated list of events with optional filters
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      - description: Sort field
        in: query
        name: sort
        type: string
      - description: Sort order (asc/desc)
        in: query
        name: order
        type: string
      - description: Filter by school ID
        in: query
        name: school_id
        type: string
      - description: Filter by district ID
        in: query
        name: district_id
        type: string
      - description: Filter by city ID
        in: query
        name: city_id
        type: string
      - description: Filter by province ID
        in: query
        name: province_id
        type: string
      - description: Filter by event type
        in: query
        name: event_type
        type: string
      - description: Filter by status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Success'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - ApiKeyAuth: []
      summary: List events
      tags:
      - Events
...
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"